Move addMonster functions over to Board
