+++ src/components/Board.js
@@ -4,6 +4,7 @@ import Scoreboard from "./Scoreboard";
 import Field from "./Field";
 import Homebase from "./Homebase";
 import Counter from "./Counter";
+import { getRandomInt } from "../helpers";
 
 import strikeSound from "../sounds/strike.wav";
 import walkSound from "../sounds/walk.wav";
@@ -39,6 +40,13 @@ const stages = [
     rateMultiplier: 1.1,
   },
 ];
+
+const directionMap = {
+  0: "up",
+  1: "left",
+  2: "right",
+  3: "down",
+};
 
 class Board extends React.Component {
   state = {
@@ -141,6 +149,64 @@ class Board extends React.Component {
         }
       }
     }
+  };
+
+  // Update state to add monster of randomColor to randomQueue of randomField
+  addMonster = (direction = "up", queueNumber = 0, colorNumber = 0) => {
+    // Update the state for the given queue to add a monster to it
+    // Make a copy of the field
+    let fields = this.state.fields;
+
+    // Add a monster to the front of it
+    fields[direction].queues[queueNumber].push({
+      type: "monster",
+      color: colorNumber,
+    });
+
+    // Update the state
+    this.setState({ fields });
+
+    // Determine if we're over the max queue length and if so, end the stage
+    if (
+      fields[direction].queues[queueNumber].length >
+      fields[direction].queueLengthLimit
+    ) {
+      this.endStage();
+    }
+
+    // If there's 1 square or fewer between the home base and the monster, go red alert
+    if (
+      fields[direction].queueLengthLimit -
+        fields[direction].queues[queueNumber].length <=
+      1
+    ) {
+      this.setState({ redAlert: true });
+    }
+
+    // Create new Monster component within the appropriate queue (this should be handled automatically)
+  };
+
+  chooseQueue = () => {
+    let fieldNumber = getRandomInt(0, 3),
+      queueNumber = getRandomInt(0, 3),
+      colorNumber = getRandomInt(0, 3);
+
+    // If queue length is 2 monsters more than any other in field:
+    // Get length of this queue
+    let field = this.state.fields[directionMap[fieldNumber]];
+
+    let fieldWouldBeUnbalanced = (allQueues, targetQueue) =>
+      field.queues.some(
+        (thisQueue) => allQueues[targetQueue].length - thisQueue.length > 1
+      );
+
+    // Get length of shortest queue
+    while (fieldWouldBeUnbalanced(field.queues, queueNumber)) {
+      queueNumber = getRandomInt(0, 3);
+    }
+    // If longest queue - shortest queue is > 2, choose a different queue
+    // queueNumber = getRandomInt(0, 3);
+    this.addMonster(directionMap[fieldNumber], queueNumber, colorNumber);
   };
 
   playSound = (soundKey, startPoint = 0, volume = 1) => {
